Fixed inclusion of helper TextFieldDateHelper into main app
TextFieldDateHelper is not included as a helper in the app because it's
not in the helpers directory. Added a workaround for this as I cannot
figure out how to include it from the Engine without moving it into
helpers directory. The problem manifests when trying to use a datetime
field, which renders the exception: undefined method `text_field_date'
for #<#<Class..."